requirements.txt
Django==5.0.3
channels==4.0.0
daphne==4.1.0
google-generativeai==0.8.5
python-dotenv==1.0.1
pypdf==5.1.0
djangorestframework==3.15.2
django-cors-headers==4.3.1
Pillow==10.1.0
torch==2.3.1+cpu
torchvision==0.18.1+cpu
timm==0.9.12
numpy>=1.26.0
opencv-python==4.8.1.78
transformers==4.36.0
scipy>=1.10.0
